+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers\Api\V1;
 
-use App\Models\User;
 use App\Models\Ticket;
 use App\Http\Filters\V1\TicketFilter;
 use App\Http\Resources\V1\TicketResource;
@@ -10,7 +9,6 @@ use App\Http\Requests\Api\V1\StoreTicketRequest;
 use App\Http\Requests\Api\V1\UpdateTicketRequest;
 use App\Http\Requests\Api\V1\ReplaceTicketRequest;
 use App\Policies\V1\TicketPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class TicketController extends ApiController
@@ -30,14 +28,12 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-        try {
-            $this->isAble('store', Ticket::class);
-
+        if($this->isAble('store', Ticket::class)) {
             return new TicketResource(Ticket::create($request->mappedAttributes()));
 
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not allowed to update that resource', 401);
         }
+
+        return $this->error('You are not allowed to update that resource', 401);
     }
 
     /**
@@ -69,16 +65,16 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             // policy
-            $this->isAble('update', $ticket);
+            if($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
+            return $this->error('You are not allowed to update that resource', 401);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not allowed to update that resource', 401);
         }
     }
 
@@ -89,11 +85,12 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             // policy
-            $this->isAble('replace', $ticket);
+            if($this->isAble('replace', $ticket)){
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
-
-            return new TicketResource($ticket);
+            return $this->error('You are not allowed to update that resource', 401);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
@@ -109,14 +106,15 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             // policy
-            $this->isAble('delete', $ticket);
+            if($this->isAble('delete', $ticket)) {
+                $ticket->delete();
+                return $this->ok('Ticket Successfully deleted');
+            }
 
-            $ticket->delete();
+            return $this->error('You are not allowed to delete that resource', 401);
 
-            return $this->ok('Ticket Successfully deleted');
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-
         }
     }
 }
