+++ app/Http/Controllers/Api/V1/UserController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\Api\V1\ReplaceUserRequest;
 use App\Http\Resources\V1\UserResource;
 use App\Http\Requests\Api\V1\StoreUserRequest;
 use App\Http\Requests\Api\V1\UpdateUserRequest;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class UserController extends ApiController
@@ -31,13 +30,11 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            $this->isAble('store', User::class);
-
+        if($this->isAble('store', User::class)) {
             return new UserResource(User::create($request->mappedAttributes()));
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not allowed to create that resource', 401);
+
         }
+        return $this->error('You are not allowed to create that resource', 401);
     }
 
     /**
@@ -62,15 +59,16 @@ class UserController extends ApiController
             $user = User::findOrFail($user_id);
 
             // policy
-            $this->isAble('update', $user);
+            if($this->isAble('update', $user)) {
+                $user->update($request->mappedAttributes());
 
-            $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-            return new UserResource($user);
+            return $this->error('You are not allowed to update that resource', 401);
+
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not allowed to update that resource', 401);
         }
     }
 
@@ -81,11 +79,13 @@ class UserController extends ApiController
             $user = User::findOrFail($user_id);
 
             // policy
-            $this->isAble('replace', $user);
+            if($this->isAble('replace', $user)) {
+                $user->update($request->mappedAttributes());
 
-            $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-            return new UserResource($user);
+            return $this->error('You are not allowed to update that resource', 401);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found.', 404);
@@ -101,11 +101,14 @@ class UserController extends ApiController
             $user = User::findOrFail($user_id);
 
             // policy
-            $this->isAble('delete', $user);
+            if($this->isAble('delete', $user)) {
+                $user->delete();
 
-            $user->delete();
+                return $this->ok('User Successfully deleted');
+            }
 
-            return $this->ok('User Successfully deleted');
+            return $this->error('You are not allowed to delete that resource', 401);
+
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found.', 404);
